requirements.txt
tensorflow>=2.9.0
opencv-python
numpy
matplotlib
pillow
scikit-learn
keras
jupyter
notebook
seaborn
